requirements.txt
gunicorn
requests
flask
mysql-connector
PyMySQL
Flask-MySQL
python-dotenv
finnhub-python